services/rag_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pydantic-settings==2.3.1

# 数据库
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
pgvector==0.3.1

# LLM
anthropic==0.28.0

# 对象存储
boto3==1.34.100
botocore==1.34.100

# 可观测
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
openinference-instrumentation-anthropic==0.1.4

# 工具
python-dotenv==1.0.1
jsonschema==4.22.0
structlog==24.2.0
